Add wp-cli
Remove todo, proxy php & ngrok works

+++ cli/ValetPlus/Binary.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace WeProvide\ValetPlus;
 
 use DomainException;
+use Valet\Brew;
 use Valet\CommandLine;
 use Valet\Filesystem;
 use function Valet\info;
@@ -18,6 +19,8 @@ class Binary
     protected const N98_MAGERUN_2 = 'magerun2';
     /** @var string */
     protected const DRUSH_LAUNCHER = 'drush';
+    /** @var string */
+    protected const WP_CLI = 'wp';
 
     /**
      * Supported binaries for the binary manager. Example:
@@ -28,7 +31,7 @@ class Binary
      *    'bin_location' => '/usr/local/bin'
      * ]
      */
-    protected const SUPPORTED_CUSTOM_BINARIES = [
+    protected const SUPPORTED_BINARIES = [
         self::N98_MAGERUN    => [
             'url'          => 'https://files.magerun.net/n98-magerun-2.3.0.phar',
             'shasum'       => 'b3e09dafccd4dd505a073c4e8789d78ea3def893cfc475a214e1154bff3aa8e4',
@@ -46,22 +49,30 @@ class Binary
             'shasum'       => '0ae18cd3f8745fdd58ab852481b89428b57be6523edf4d841ebef198c40271be',
             'bin_location' => BREW_PREFIX . '/bin/',
             'framework'    => 'Drupal'
+        ],
+        self::WP_CLI         => [
+            'brew_formula' => 'wp-cli'
         ]
     ];
 
+    /** @var Brew */
+    protected $brew;
     /** @var CommandLine */
     protected $cli;
     /** @var Filesystem */
     protected $files;
 
     /**
+     * @param Brew $brew
      * @param CommandLine $cli
      * @param Filesystem $files
      */
     public function __construct(
+        Brew        $brew,
         CommandLine $cli,
         Filesystem  $files
     ) {
+        $this->brew  = $brew;
         $this->cli   = $cli;
         $this->files = $files;
     }
@@ -73,7 +84,7 @@ class Binary
      */
     public function getSupported()
     {
-        return array_keys(static::SUPPORTED_CUSTOM_BINARIES);
+        return array_keys(static::SUPPORTED_BINARIES);
     }
 
     /**
@@ -82,9 +93,17 @@ class Binary
      * @param string $binary
      * @return bool
      */
-    public function installed($binary)
+    public function installed($binary): bool
     {
-        return $this->files->exists(static::SUPPORTED_CUSTOM_BINARIES[$binary]['bin_location'] . $binary);
+        if (isset(static::SUPPORTED_BINARIES[$binary]['bin_location'])) {
+            return $this->files->exists(static::SUPPORTED_BINARIES[$binary]['bin_location'] . $binary);
+        }
+
+        if (isset(static::SUPPORTED_BINARIES[$binary]['brew_formula'])) {
+            return $this->brew->installed(static::SUPPORTED_BINARIES[$binary]['brew_formula']);
+        }
+
+        return false;
     }
 
     /**
@@ -92,10 +111,8 @@ class Binary
      */
     public function install()
     {
-        foreach (static::SUPPORTED_CUSTOM_BINARIES as $binary => $binInfo) {
-            if (!$this->installed($binary)) {
-                $this->installBinary($binary);
-            }
+        foreach (static::SUPPORTED_BINARIES as $binary => $binInfo) {
+            $this->installBinary($binary);
         }
     }
 
@@ -106,7 +123,7 @@ class Binary
      */
     public function installBinary($binary)
     {
-        if (!isset(static::SUPPORTED_CUSTOM_BINARIES[$binary])) {
+        if (!isset(static::SUPPORTED_BINARIES[$binary])) {
             throw new DomainException(
                 sprintf(
                     'Invalid binary given. Available binaries: %s',
@@ -120,29 +137,38 @@ class Binary
             return;
         }
 
-        $url      = $this->getUrl($binary);
-        $urlSplit = explode('/', $url);
-        $fileName = $urlSplit[count($urlSplit) - 1];
+        // Download and install binary.
+        if (isset(static::SUPPORTED_BINARIES[$binary]['bin_location'])) {
+            $url      = $this->getUrl($binary);
+            $urlSplit = explode('/', $url);
+            $fileName = $urlSplit[count($urlSplit) - 1];
 
-        // Download binary file.
-        info("Binary $binary installing from: $url");
-        $this->cli->passthru("cd /tmp && curl -OL $url");
+            // Download binary file.
+            info("Binary $binary installing from: $url");
+            $this->cli->passthru("cd /tmp && curl -OL $url");
 
-        // Check the checksum of downloaded file.
-        if (!$this->checkShasum($binary, $fileName)) {
-            $this->cli->runAsUser("rm /tmp/$fileName");
-            warning("Binary $binary could not be installed, $fileName checksum does not match: " . $this->getShasum($binary));
+            // Check the checksum of downloaded file.
+            if (!$this->checkShasum($binary, $fileName)) {
+                $this->cli->runAsUser("rm /tmp/$fileName");
+                warning("Binary $binary could not be installed, $fileName checksum does not match: " . $this->getShasum($binary));
 
-            return;
+                return;
+            }
+
+            // Move file.
+            $binLocation = $this->getBinLocation($binary);
+            $this->cli->run("sudo mv /tmp/$fileName $binLocation");
+
+            // Make file executable.
+            $this->cli->run("sudo chmod +x $binLocation");
+            info("Binary $binary installed to: $binLocation");
         }
 
-        // Move file.
-        $binLocation = $this->getBinLocation($binary);
-        $this->cli->run("sudo mv /tmp/$fileName $binLocation");
-
-        // Make file executable.
-        $this->cli->run("sudo chmod +x $binLocation");
-        info("Binary $binary installed to: $binLocation");
+        // Install brew formula.
+        if (isset(static::SUPPORTED_BINARIES[$binary]['brew_formula'])) {
+            $formula = static::SUPPORTED_BINARIES[$binary]['brew_formula'];
+            $this->brew->ensureInstalled($formula);
+        }
     }
 
     /**
@@ -150,7 +176,7 @@ class Binary
      */
     public function uninstall()
     {
-        foreach (static::SUPPORTED_CUSTOM_BINARIES as $binary => $binInfo) {
+        foreach (static::SUPPORTED_BINARIES as $binary => $binInfo) {
             if ($this->installed($binary)) {
                 $this->uninstallBinary($binary);
             }
@@ -164,12 +190,21 @@ class Binary
      */
     public function uninstallBinary($binary)
     {
-        $binaryLocation = $this->getBinLocation($binary);
-        $this->cli->runAsUser('rm ' . $binaryLocation);
-        if ($this->files->exists($binaryLocation)) {
-            throw new DomainException('Could not remove binary! Please remove manually using: rm ' . $binaryLocation);
+        // Remove downloaded binary.
+        if (isset(static::SUPPORTED_BINARIES[$binary]['bin_location'])) {
+            $binaryLocation = $this->getBinLocation($binary);
+            $this->cli->runAsUser('rm ' . $binaryLocation);
+            if ($this->files->exists($binaryLocation)) {
+                throw new DomainException('Could not remove binary! Please remove manually using: rm ' . $binaryLocation);
+            }
+            info("Binary $binary successfully uninstalled!");
         }
-        info("Binary $binary successfully uninstalled!");
+
+        // Uninstall brew formula.
+        if (isset(static::SUPPORTED_BINARIES[$binary]['brew_formula'])) {
+            $formula = static::SUPPORTED_BINARIES[$binary]['brew_formula'];
+            $this->brew->uninstallFormula($formula);
+        }
     }
 
     /**
@@ -177,8 +212,11 @@ class Binary
      */
     protected function update($binary)
     {
-        $binLocation = $this->getBinLocation($binary);
-        $this->cli->run("sudo $binLocation self-update");
+        if (isset(static::SUPPORTED_BINARIES[$binary]['bin_location'])) {
+            info("Binary $binary updating...");
+            $binLocation = $this->getBinLocation($binary);
+            $this->cli->run("sudo $binLocation self-update");
+        }
     }
 
     /**
@@ -189,8 +227,8 @@ class Binary
      */
     protected function getUrl($binary)
     {
-        if (array_key_exists('url', static::SUPPORTED_CUSTOM_BINARIES[$binary])) {
-            return static::SUPPORTED_CUSTOM_BINARIES[$binary]['url'];
+        if (array_key_exists('url', static::SUPPORTED_BINARIES[$binary])) {
+            return static::SUPPORTED_BINARIES[$binary]['url'];
         }
         throw new DomainException('url key is required for binaries.');
     }
@@ -203,8 +241,8 @@ class Binary
      */
     protected function getShasum($binary)
     {
-        if (array_key_exists('shasum', static::SUPPORTED_CUSTOM_BINARIES[$binary])) {
-            return static::SUPPORTED_CUSTOM_BINARIES[$binary]['shasum'];
+        if (array_key_exists('shasum', static::SUPPORTED_BINARIES[$binary])) {
+            return static::SUPPORTED_BINARIES[$binary]['shasum'];
         }
         throw new DomainException('shasum key is required for binaries.');
     }
@@ -217,8 +255,8 @@ class Binary
      */
     protected function getBinLocation($binary)
     {
-        if (array_key_exists('bin_location', static::SUPPORTED_CUSTOM_BINARIES[$binary])) {
-            return static::SUPPORTED_CUSTOM_BINARIES[$binary]['bin_location'] . $binary;
+        if (array_key_exists('bin_location', static::SUPPORTED_BINARIES[$binary])) {
+            return static::SUPPORTED_BINARIES[$binary]['bin_location'] . $binary;
         }
         throw new DomainException('bin_location key is required for binaries.');
     }
